add first step of aggregator + clean trash in .c (also from freccia branch)

+++ srcs/tcaps/tcaps_cut_paste.c
@@ -7,7 +7,7 @@ static void		tcaps_ctrl_k(t_env *e)
 	char	*tmp;
 
 	j = -1;
-	if (NB_MOVE >= ft_strlen(e->line))
+	if (NB_MOVE >= (int)ft_strlen(e->line))
 		return ;
 	strfree(&e->cut);
 	e->cut = ft_strdup(&e->line[NB_MOVE]);
